requirements.txt
feedparser==6.*
notion-client==2.*
python-dateutil>=2.8,<3
trafilatura==2.*
beautifulsoup4>=4.12,<5
lxml>=5.2,<6
